refactor: deconstruct SimpleMintingComponents.tsx into molecules
