+++ eslint.config.mjs
@@ -341,7 +341,8 @@ export default ts.config(
       "src/components/organisms/PromptBubbleEditor.tsx",
       "src/components/organisms/PromptDeclutterControls.tsx",
       "src/components/organisms/ShareCardModal.tsx",
-      "src/components/organisms/SimpleMintingComponents.tsx",
+      "src/components/molecules/simple-minting/KeywordSuggestions.tsx",
+      "src/components/molecules/simple-minting/ThumbnailPreview.tsx",
       "src/components/organisms/SimpleWorkbenchModal.tsx",
       "src/components/organisms/StorageManagerSection.tsx",
       "src/components/organisms/SyncStatusMessage.tsx",
